% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rtf_multiple_specs.R
\name{create_rtf_multiple_specs}
\alias{create_rtf_multiple_specs}
\title{Create RTF Documents Based on Multiple Specifications}
\usage{
create_rtf_multiple_specs(
  data,
  split_list,
  specs_list = NULL,
  just_list = NULL
)
}
\arguments{
\item{data}{A data frame, the dataset to be used in the RTF document.}

\item{split_list}{List, defining the columns for each split of data.}

\item{specs_list}{List of lists, providing specifications for each subset of data. Each element of the list corresponds to a subset defined in \code{split_list}.}

\item{just_list}{List of vectors, defining the justification for each column in each subset of data.}
}
\value{
A character vector, where each element represents a line in the final RTF document.
}
\description{
This function generates RTF documents with tables based on multiple specifications provided.
It splits the data, applies the specifications, and then combines the output.
}
\examples{
\dontrun{
# Simple Example Data
data_simple <- data.frame(
  label = c('', 'A', 'B', '', ''),
  col1 = c('', '10 (20.0)', '20 (40.0)', '$NEWPAGE$', '$NEWPAGE$'),
  col2 = c('', '5 (10.0)', '15 (30.0)', '', '')
)

# Specifying Splits and Specifications
split_list_simple <- list(c(1, 2), 1, 2)
specs_list_simple <- list(
  list(list(label='Category', just='l', pct=50), list(label='Value (\%)', just='c', pct=50)),
  list(list(label='Category', just='l', pct=50), list(label='Value', just='c', pct=50))
)
just_list_simple <- list(c('l', 'c'), c('l', 'c'))

# Creating RTF Document
out_rtf_txt_simple <- create_rtf_multiple_specs(
  data = data_simple,
  split_list = split_list_simple,
  specs_list = specs_list_simple,
  just_list = just_list_simple
)
}
}
