% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rtf_table.R
\name{create_rtf_table}
\alias{create_rtf_table}
\title{Create RTF Table}
\usage{
create_rtf_table(
  data,
  specs = NULL,
  col_just = NULL,
  header_text_lines = NULL,
  footer_text_lines = NULL,
  input_status = "qa",
  output_status = "qa",
  PAGE_WIDTH = 133,
  PAGE_HEIGHT = 45
)
}
\arguments{
\item{data}{A data frame containing the data to be included in the table.}

\item{specs}{A list of specifications for the table header.}

\item{col_just}{A character vector specifying the justification of each column.}

\item{header_text_lines}{A character vector containing the lines of text to be
included in the table header.}

\item{footer_text_lines}{A character vector containing the lines of text to be
included in the table footer.}

\item{input_status}{A string that represents the source environment of the input data.
Can be either 'qa', or 'prd'. Default is 'qa'.}

\item{output_status}{A string that represents in which environment will the output be saved.
Can be either 'qa', or 'prd'. Default is 'qa'.}

\item{PAGE_WIDTH}{The width of the page. Default is 133.}

\item{PAGE_HEIGHT}{The height of the page. Default is 45.}
}
\value{
A character vector containing the RTF code for the table.
}
\description{
This function creates a table in RTF format from the given data, specifications,
header text, and footer text.
}
\details{
The function creates a table in RTF format with the specified header, body, and footer.
The `data` parameter is a data frame containing the data to be included in the table body.
The `specs` parameter is a list of specifications for the table header.
The `col_just` parameter is a character vector specifying the justification of each column.
The `header_text_lines` parameter is a character vector containing the lines of text to be included in the table header.
The `footer_text_lines` parameter is a character vector containing the lines of text to be included in the table footer.
The `input_status` parameter is a specify whether the data used in the table is in development (='qa'), or production (='prd') environment.
The `output_status` parameter is a specify whether the table is saved in development (='qa'), or production (='prd') environment.
The `PAGE_WIDTH` parameter specifies the width of the page.
The `PAGE_HEIGHT` parameter specifies the height of the page.

If $NEWPAGE$ is included in any of the table body rows, it will start a new page.
}
\examples{
\dontrun{
#' data <- data.frame(
  col1 = c("The quick brown fox",
           "$NEWPAGE$",
           "jumps over the lazy dog",
           "$NEWPAGE$",
           "She sells sea shells down by the sea shore\nnew line"),
  col2 = c("Lorem ipsum dolor sit amet",
           "",
           "consectetur adipiscing elit",
           "",
           "Sepmre fidelis")
)

specs <- list(
  list(label='Quotes', just='c',
       sub=list(
         list(label='English', just='c'),
         list(label='Latin', just='l')
       ))
)

header_text <- c(
  "this is my header this is my header this is my header this is my header",
  "Safety Population"
)

footer_text <- c(
  "this is my footer this is my footer this is my footer this is my footer",
  "",
  "a path"
)

out_rtf <- create_rtf_table(data = data, header_text_lines = header_text,
                            footer_text_lines = footer_text, specs = specs)
writeLines(out_rtf, 'tmp_tfl.rtf')
}

}
\seealso{
\code{\link{create_header_from_specs}} for the detailed explanation of the `specs` parameter.
}
