% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_body.R
\name{create_table_body}
\alias{create_table_body}
\title{Create Table Body}
\usage{
create_table_body(
  data,
  table_width,
  widths = NULL,
  positions = NULL,
  just = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to be included in the table body.}

\item{table_width}{An integer specifying the total width of the table.}

\item{widths}{(Optional) A vector of integers specifying the width of each column.
If NULL or if its length is less than the number of columns in the data, it will be calculated.}

\item{positions}{(Optional) A vector of integers specifying the starting position of each column.
If NULL or if its length is less than the number of columns in the data, it will be calculated.}

\item{just}{(Optional) A vector of characters specifying the text justification for each column.
It can be 'l' for left-justified, 'c' for center-justified, or 'r' for right-justified.
If NULL or if its length is less than the number of columns in the data,
all columns will be left-justified.}
}
\value{
A vector of character strings, each representing a line (row) of the table body.
}
\description{
This function creates the body of a table from a given data frame. The table body is created
based on the provided column widths, positions, and justification. If these parameters are not
provided, they are calculated based on the table width and the number of columns.
}
\examples{
data <- data.frame(
  col1 = c("The quick brown fox", "", "jumps over the lazy dog"),
  col2 = c("Lorem ipsum dolor sit amet", "", "consectetur adipiscing elit")
)

table_width <- 40
widths <- c(10, 20)
positions <- c(1, 13)
just <- c('c', 'l')

create_table_body(data, table_width, widths, positions, just)

}
\seealso{
\code{\link{create_header_from_specs}}
}
