% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\docType{data}
\name{Header}
\alias{Header}
\title{Header ggproto object}
\format{
An object of class \code{Header} (inherits from \code{ggproto}, \code{gg}) of length 12.
}
\usage{
Header
}
\description{
The Header object has several methods:

\describe{
  \item{add_line}{Adds a line of text to the header. Takes a single argument,
  \code{txt}, a character string containing the text to be added.}
  \item{get_lines}{Returns the lines of text in the header.}
  \item{new}{Creates a new Header object.}
  \item{add_empty_line}{Adds an empty line to the header.}
  \item{set_data_status}{Sets the input and output status of the header. Takes two arguments,
  \code{instat}, a character string indicating the input status, and \code{outstat}, a character
  string indicating the output status.}
  \item{complete_header}{Completes the header by adding the current date and time,
  and the input and output status.}
  \item{get_depth}{Returns the number of lines in the header.}
}
}
\details{
A ggproto object for creating and managing table headers.
}
\section{Fields}{

\describe{
\item{\code{width}}{The total width of the header. Default is 133.}

\item{\code{lines}}{A character vector containing the lines of text in the header.}

\item{\code{current_datetime}}{A character string containing the current date and time.
It is updated each time a new Header object is created.}

\item{\code{input_status}}{A character string indicating the input status. Default is 'qa'.}

\item{\code{output_status}}{A character string indicating the output status. Default is 'qa'.}
}}

\examples{
new_header <- Header$new()
new_header$add_line("this is my header this is my header this is my header this is my header")
new_header$add_line("Safety Population")
new_header$complete_header()
new_header$get_depth()
for (l in new_header$get_lines()) {
  cat(l, '\n')
}

}
\keyword{datasets}
