% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{justify_text}
\alias{justify_text}
\title{Justify Text}
\usage{
justify_text(text, width, justification)
}
\arguments{
\item{text}{The text to be justified.}

\item{width}{The width of the field.}

\item{justification}{The justification type: 'c' for center, 'r' for right, and 'l' for left.}
}
\value{
The justified text.
}
\description{
This function justifies a text to the left, center, or right.
}
\examples{
\dontrun{justify_text("hello", 10, 'c')}
# [1] "   hello   "
}
