% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_text}
\alias{split_text}
\title{Split Text Into Lines}
\usage{
split_text(text, width)
}
\arguments{
\item{text}{The text to be split.}

\item{width}{The maximum width of a line.}
}
\value{
A list of lines.
}
\description{
This function splits a text into lines of a specified width.
}
\examples{
\dontrun{split_text('lorem bla baglk asdf asdfasfd  fdsafasdf', 15)}
# [[1]]
# [1] "lorem bla baglk "
#
# [[2]]
# [1] "asdf asdfasfd  "
#
# [[3]]
# [1] "fdsafasdf "
}
