% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substr_replace}
\alias{substr_replace}
\title{Replace a Substring Within a String}
\usage{
substr_replace(main_str, replace_str, start, stop)
}
\arguments{
\item{main_str}{The main string.}

\item{replace_str}{The string to replace the substring with.}

\item{start}{The starting index of the substring to be replaced.}

\item{stop}{The ending index of the substring to be replaced.}
}
\value{
The modified string.
}
\description{
This function replaces a substring within a string.
}
\examples{
\dontrun{substr_replace("hello world", "new", 7, 11)}
# [1] "hello new"
}
